% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSfunctions.R
\name{get.risk}
\alias{get.risk}
\title{Calculate the proportion of vector elements above a threshold (or below or any other operator) to estimate risk.}
\usage{
get.risk(vec, thres, op = ">", na.rm = FALSE)
}
\arguments{
\item{vec}{vector}

\item{thres}{treshold value}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Vector of interpolated values
}
\description{
\code{get.risk} returns the proportion of vector values below or above a threshold.
}
