% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSfunctions.R
\name{getColors}
\alias{getColors}
\title{Define colour scheme.}
\usage{
getColors(n, levels = NA, colSc = "normal")
}
\arguments{
\item{n}{Number of colour classes, used if levels are NA.}

\item{levels}{Levels defining the classes, if NA.}

\item{colSc}{Colour scheme: [normal|normal_reversed|normalwhite|sharp|sharp100|sharp_redgreenblue|sharp_cyanbrown|any of the RColorBrewer color palettes|vector of colours]; defaults to normal which uses topo.colors(); sharp* requires levels to be set that cross 0 (100); if colSc is not any of the pre-defined strings for colour schemes, colSc is returned and assumed to be a vector that defines colours}
}
\value{
Vector of colours.
}
\description{
\code{getColors} returns a vector with colours.
}
