% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSfunctions.R
\name{interp.IRS}
\alias{interp.IRS}
\title{Interpolate points on an impact response surface}
\usage{
interp.IRS(dat, xvar = "X", yvar = "Y", zvar = "Z", xo, yo, ...)
}
\arguments{
\item{dat}{data.table object with 3 columns (x, y, z)}

\item{xvar}{Character vector containing the column name for the x dimension}

\item{yvar}{Character vector containing the column name for the y dimension}

\item{zvar}{Character vector containing the column name for the z dimension}

\item{xo}{vector of x coordinates defining points to interpolate to (same length as yo)}

\item{yo}{vector of y coordinates defining points to interpolate to (same length as xo)}

\item{...}{additional arguments for akima::interpp()}
}
\value{
Vector of interpolated values
}
\description{
\code{interpIRS} returns values interpolated from an impact response surface
}
