% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSfunctions.R
\name{plotIRS}
\alias{plotIRS}
\title{Plot a single IRS.}
\usage{
plotIRS(
  dat,
  xvar = "deltaT",
  yvar = "Pchange",
  xlab = expression(paste(plain("Temperature change ("), degree, plain("C)"))),
  ylab = "Precipitation change (\%)",
  var = "grain_DM",
  add = FALSE,
  col = "black",
  lty = 1,
  lwd = 1,
  title = NA,
  levels = NULL,
  cex = 1,
  labcex = 0.6,
  drawlabels = TRUE,
  colSc = "normal",
  xlim = NULL,
  ylim = NULL,
  xaxt = "s",
  yaxt = "s",
  legend = FALSE,
  contours = TRUE,
  stippled = TRUE
)
}
\arguments{
\item{dat}{data.table object that contains three columns for x-, y- and z-dimension}

\item{xvar}{Column name for the x-axis}

\item{yvar}{Column name for the y-axis}

\item{xlab}{x-axis label text}

\item{ylab}{y-axis label text}

\item{add}{if FALSE, starts a new plot, if TRUE, plots ontop on the current device}

\item{col}{colour of contour lines, if NA, no contours are plotted}

\item{lty}{line type of contour lines}

\item{lwd}{line width of contour lines}

\item{title}{plot title}

\item{levels}{levels for contour lines}

\item{cex}{font size}

\item{labcex}{font size for contour labels}

\item{colSc}{colour scheme defined in \code{\link{getColors}} or NULL if no filled contours should be plot}

\item{xlim}{Ranges of the x- and y-axes to plot (defaults to the range specified in dat$xvar and dat$yvar)}

\item{ylim}{Ranges of the x- and y-axes to plot (defaults to the range specified in dat$xvar and dat$yvar)}

\item{legend}{if true adds a legend to the plot}

\item{contours}{if true plots data as contours; if false, plots the data as a grid; can also be used with transparent colour plotted on top of filled contour plots,    e.g. \code{plotIRSnoaxes(dat=myDat, var="Z5", colSc=rgb(0,0,0, alpha=.2), levels=c(.5,1.5), contours=FALSE, add=TRUE)}}

\item{stippled}{if contours==FALSE, plot selected grid cells as stippled or hatched area}

\item{varn}{Column name for the z-axis}
}
\value{
Vector of the levels used for the plot.
}
\description{
\code{plotIRS} no return values.
}
