% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSfunctions.R
\name{plotIRSnoaxes}
\alias{plotIRSnoaxes}
\title{Plot a single IRS without axes titles, wrapper for \code{\link{plotIRS}}}
\usage{
plotIRSnoaxes(...)
}
\arguments{
\item{...}{Arguments passed over to \code{\link{plotIRS}}.}
}
\value{
Vector of the levels is for the plot.
}
\description{
\code{plotIRSnoaxes} no return values.
}
